chore: rename ctx and add eslint disable lines
