+++ contexts/index.ts
@@ -1,3 +1,3 @@
 export * from "./SiteMinting"
 export * from "./Refresh"
-export * from "./MongoUri"
+export * from "./HasMongoUri"
